refactor: View record + Projection 인터페이스 dto 패키지 분리 (#74)
