+++ core/src/main/java/com/mailsangja/core/dto/inbox/ThreadDetailResult.java
@@ -2,7 +2,7 @@ package com.mailsangja.core.dto.inbox;
 
 import com.mailsangja.db.entity.mail.Message;
 import com.mailsangja.db.entity.mail.Thread;
-import com.mailsangja.db.port.MessageLabelView;
+import com.mailsangja.db.dto.MessageLabelView;
 
 import java.util.List;
 import java.util.Map;
