+++ core/src/main/java/com/mailsangja/core/dto/trash/TrashMessageItemResponse.java
@@ -2,7 +2,7 @@ package com.mailsangja.core.dto.trash;
 
 import com.mailsangja.db.entity.mail.Direction;
 import com.mailsangja.db.entity.mail.Message;
-import com.mailsangja.db.port.MessageLabelView;
+import com.mailsangja.db.dto.MessageLabelView;
 import io.swagger.v3.oas.annotations.media.Schema;
 
 import java.time.LocalDateTime;
