+++ core/src/main/java/com/mailsangja/core/dto/trash/TrashThreadDetailResponse.java
@@ -3,7 +3,7 @@ package com.mailsangja.core.dto.trash;
 import com.mailsangja.core.dto.inbox.MessageResponse;
 import com.mailsangja.db.entity.mail.Message;
 import com.mailsangja.db.entity.mail.Thread;
-import com.mailsangja.db.port.MessageLabelView;
+import com.mailsangja.db.dto.MessageLabelView;
 import io.swagger.v3.oas.annotations.media.Schema;
 
 import java.time.LocalDateTime;
