+++ core/src/main/java/com/mailsangja/core/dto/trash/TrashThreadSummaryResponse.java
@@ -4,7 +4,7 @@ import com.mailsangja.core.dto.inbox.AttachmentResponse;
 import com.mailsangja.core.dto.inbox.MailAddressResponse;
 import com.mailsangja.db.entity.mail.Attachment;
 import com.mailsangja.db.entity.mail.Thread;
-import com.mailsangja.db.port.ThreadLabelView;
+import com.mailsangja.db.dto.ThreadLabelView;
 import io.swagger.v3.oas.annotations.media.Schema;
 
 import java.time.LocalDateTime;
