+++ core/src/main/java/com/mailsangja/core/facade/TrashFacade.java
@@ -16,8 +16,8 @@ import com.mailsangja.db.entity.mail.MailAccount;
 import com.mailsangja.db.entity.mail.Message;
 import com.mailsangja.db.entity.mail.Thread;
 import com.mailsangja.db.entity.user.User;
-import com.mailsangja.db.port.MessageLabelView;
-import com.mailsangja.db.port.ThreadLabelView;
+import com.mailsangja.db.dto.MessageLabelView;
+import com.mailsangja.db.dto.ThreadLabelView;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Slice;
 import org.springframework.stereotype.Component;
