+++ core/src/main/java/com/mailsangja/core/service/inbox/InboxQueryService.java
@@ -10,8 +10,8 @@ import com.mailsangja.db.entity.mail.Message;
 import com.mailsangja.db.entity.mail.Thread;
 import com.mailsangja.db.port.ContactRepositoryPort;
 import com.mailsangja.db.port.MessageRepositoryPort;
-import com.mailsangja.db.port.MessageLabelView;
-import com.mailsangja.db.port.ThreadLabelView;
+import com.mailsangja.db.dto.MessageLabelView;
+import com.mailsangja.db.dto.ThreadLabelView;
 import com.mailsangja.db.port.ThreadRepositoryPort;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Pageable;
