+++ db/src/main/java/com/mailsangja/db/port/MessageRepositoryPort.java
@@ -1,5 +1,7 @@
 package com.mailsangja.db.port;
 
+import com.mailsangja.db.dto.MessageLabelView;
+import com.mailsangja.db.dto.ThreadLabelView;
 import com.mailsangja.db.entity.mail.Message;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Slice;
